refactor: split ExportTab and extract useSnapshotPanel hook
ExportConfirm moves to its own component, and the format extension/
description maps join the rest of the format metadata in types.ts.
SnapshotPanel's state and handlers (create/preview/delete/rename/
compare) move into a useSnapshotPanel hook, following the same
hook-holds-state, component-just-renders pattern already used by
useStaged and useSortableListEditor.
